+++ tests/api/test_order.py
@@ -489,7 +489,7 @@ def test_payment_information_order_events_query(
     assert data["paymentGateway"] == payment_dummy.gateway
 
 
-def test_non_staff_user_can_only_see_his_order(user_api_client, order):
+def test_non_staff_user_cannot_only_see_his_order(user_api_client, order):
     query = """
     query OrderQuery($id: ID!) {
         order(id: $id) {
@@ -500,16 +500,27 @@ def test_non_staff_user_can_only_see_his_order(user_api_client, order):
     ID = graphene.Node.to_global_id("Order", order.id)
     variables = {"id": ID}
     response = user_api_client.post_graphql(query, variables)
-    content = get_graphql_content(response)
-    order_data = content["data"]["order"]
-    assert order_data["number"] == str(order.pk)
+    assert_no_permission(response)
 
-    order.user = None
-    order.save()
-    response = user_api_client.post_graphql(query, variables)
+
+def test_query_order_as_service_account(
+    service_account_api_client, permission_manage_orders, order
+):
+    query = """
+    query OrderQuery($id: ID!) {
+        order(id: $id) {
+            token
+        }
+    }
+    """
+    ID = graphene.Node.to_global_id("Order", order.id)
+    variables = {"id": ID}
+    response = service_account_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
     content = get_graphql_content(response)
     order_data = content["data"]["order"]
-    assert not order_data
+    assert order_data["token"] == order.token
 
 
 def test_draft_order_create(
@@ -2016,6 +2027,36 @@ def test_authorized_access_to_order_by_token(
     assert content["data"]["orderByToken"]["user"]["id"] == customer_user_id
 
 
+def test_query_draft_order_by_token_with_requester_as_customer(
+    user_api_client, draft_order
+):
+    draft_order.user = user_api_client.user
+    draft_order.save(update_fields=["user"])
+    query = """
+    query OrderByToken($token: UUID!) {
+        orderByToken(token: $token) {
+            id
+        }
+    }
+    """
+    response = user_api_client.post_graphql(query, {"token": draft_order.token})
+    content = get_graphql_content(response)
+    assert not content["data"]["orderByToken"]
+
+
+def test_query_draft_order_by_token_as_anonymous_customer(api_client, draft_order):
+    query = """
+    query OrderByToken($token: UUID!) {
+        orderByToken(token: $token) {
+            id
+        }
+    }
+    """
+    response = api_client.post_graphql(query, {"token": draft_order.token})
+    content = get_graphql_content(response)
+    assert not content["data"]["orderByToken"]
+
+
 MUTATION_CANCEL_ORDERS = """
     mutation CancelManyOrders($ids: [ID]!, $restock: Boolean!) {
         orderBulkCancel(ids: $ids, restock: $restock) {
@@ -2618,7 +2659,7 @@ def test_user_does_not_need_permission_to_update_meta(
     errors = content["data"]["orderUpdateMeta"]["errors"]
     assert len(errors) == 0
     order.refresh_from_db()
-    assert order.get_meta(namespace="test", client="client1") == {"foo": "bar"}
+    assert order.get_value_from_metadata("foo") == "bar"
 
 
 def test_user_with_permission_can_update_private_meta(
@@ -2638,7 +2679,7 @@ def test_user_with_permission_can_update_private_meta(
     errors = content["data"]["orderUpdatePrivateMeta"]["errors"]
     assert len(errors) == 0
     order.refresh_from_db()
-    assert order.get_private_meta(namespace="test", client="client1") == {"foo": "bar"}
+    assert order.get_value_from_private_metadata("foo") == "bar"
 
 
 @pytest.fixture
@@ -2683,17 +2724,15 @@ def clear_meta_private_variables(order):
         "input": {
             "namespace": "testPrivate",
             "clientName": "clientPrivate1",
-            "key": "foo",
+            "key": "foofoo",
         },
     }
 
 
 @pytest.fixture
 def order_with_meta(order):
-    order.store_meta(namespace="test", client="client1", item={"foo": "bar"})
-    order.store_private_meta(
-        namespace="testPrivate", client="clientPrivate1", item={"foo": "bar"}
-    )
+    order.store_value_in_metadata(items={"foo": "bar"})
+    order.store_value_in_private_metadata(items={"foofoo": "barbar"})
     order.save()
     return order
 
@@ -2732,8 +2771,7 @@ def test_user_with_permission_can_clear_meta(
     errors = content["data"]["orderClearMeta"]["errors"]
     assert len(errors) == 0
     order_with_meta.refresh_from_db()
-    current_meta = order_with_meta.get_meta(namespace="test", client="client1")
-    assert current_meta == {}
+    assert not order_with_meta.get_value_from_metadata("foo")
 
 
 def test_user_with_permission_can_clear_private_meta(
@@ -2754,5 +2792,4 @@ def test_user_with_permission_can_clear_private_meta(
     errors = content["data"]["orderClearPrivateMeta"]["errors"]
     assert len(errors) == 0
     order_with_meta.refresh_from_db()
-    current_meta = order_with_meta.get_private_meta(namespace="test", client="client1")
-    assert current_meta == {}
+    assert not order_with_meta.get_value_from_private_metadata(key="foofoo")