+++ tests/api/test_view.py
@@ -5,6 +5,7 @@ import graphene
 import pytest
 from django.test import override_settings
 
+from saleor.demo.views import EXAMPLE_QUERY
 from saleor.graphql.product.types import Product
 from saleor.graphql.views import handled_errors_logger, unhandled_errors_logger
 
@@ -52,16 +53,11 @@ def test_batch_queries(category, product, api_client):
     assert data["category"]["name"] == category.name
 
 
-def test_graphql_view_get_in_non_debug_mode(client):
+@pytest.mark.parametrize("playground_on, status", [(True, 200), (False, 405)])
+def test_graphql_view_get_enabled_or_disabled(client, settings, playground_on, status):
+    settings.PLAYGROUND_ENABLED = playground_on
     response = client.get(API_PATH)
-    assert response.status_code == 405
-
-
-@override_settings(DEBUG=True)
-def test_graphql_view_get_in_debug_mode(client):
-    response = client.get(API_PATH)
-    assert response.status_code == 200
-    assert response.templates[0].name == "graphql/playground.html"
+    assert response.status_code == status
 
 
 def test_graphql_view_options(client):
@@ -232,3 +228,9 @@ def test_unexpected_exceptions_are_logged_in_their_own_logger(
     assert graphql_log_handler.messages == [
         "saleor.graphql.errors.unhandled[ERROR].NotImplementedError"
     ]
+
+
+def test_example_query(api_client, product):
+    response = api_client.post_graphql(EXAMPLE_QUERY)
+    content = get_graphql_content(response)
+    assert content["data"]["products"]["edges"][0]["node"]["name"] == product.name