+++ CHANGELOG.md
@@ -64,8 +64,18 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add slug to Warehouse, Product, ProductType and update slug in models which already using it - #5196 by @IKarbowiak
 - Add mutation for assigning, unassigning shipping zones to warehouse - #5217 by @kswiatek92
 - Fix passing addresses to `PaymentData` objects - #5223 by @maarcingebala
+- Return `null` when querying `me` as an anonymous user - #5231 by @maarcingebala
+- Added `PLAYGROUND_ENABLED` environment variable/setting to allow to enable the GraphQL playground when `DEBUG` is disabled - #5254 by @NyanKiyoshi
+- Fix access to order query when request from service account - #5258 by @fowczarek
+- Customer shouldn't be able to see draft orders by token - #5259 by @fowczarek
+- Customer shouldn't be able to query checkout with another customer - #5268 by @fowczarek
+- Added integration support of Jaeger Tracing - #5282 by @NyanKiyoshi
+- Customer shouldn't be able to see draft orders by token  - #5259 by @fowczarek
 - Return `null` when querying `me` as an anonymous user - #5231 as @maarcingebala
 - Add `fulfillment created` webhook - @szewczykmira
+- Unify saleor metadata - #5178 by @fowczarek
+- Add compiled versions of emails to the repository - #5260 by @tomaszszymanski129
+
 
 ## 2.9.0
 
@@ -103,10 +113,11 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix fetching staff user without `manage_users` permission - #4835 by @fowczarek
 - Ensure that a GraphQL query is a string - #4836 by @nix010
 - Add ability to configure the password reset link - #4863 by @fowczarek
+- Fixed a performance issue where Saleor would sometimes run huge, unneeded prefetches when resolving categories or collections - #5291 by @NyanKiyoshi
 
 ### Core
 
-- Add enterprise-grade attributes management - #4351 by @dominik-zeglen and @NyanKiyoshix
+- Add enterprise-grade attributes management - #4351 by @dominik-zeglen and @NyanKiyoshi
 - Add extensions manager - #4497 by @korycins
 - Add service accounts - backend support - #4689 by @korycins
 - Add support for webhooks - #4731 by @korycins